No hybas filters just sorting
During my testing I started by removing the complexity that the hybas filters create and this is the result:

It is clear that creating multiple layers to create the hybas hierarchy results in semi-random labeling by MapBox. 
With hybas sorting the best we can do is https://github.com/kartverket/den-norske-los-kart/pull/15 
So Bente what would you like: semi-random but according to the hybas hierarchy, or just smallest numbers on top?